First changes
Command does not work properly. Also, I still have to make changes to a function, so that it is called updateDisplay that takes a  parameter called message. This function should add the message to the textarea (it's done, just change the names). +Visually indicate which of the directional buttons are valid.